+++ packages/cli/src/zosfiles/upload/ftu/FileToUSS.handler.ts
@@ -43,6 +43,8 @@ export default class FileToUSSHandler extends ZosFilesBaseHandler {
             task: task,
             responseTimeout: commandParameters.arguments.responseTimeout,
             includeHidden: commandParameters.arguments.includeHidden,
+            makeDir: commandParameters.arguments.makeDir,
+            skipDirectoryCheck: commandParameters.arguments.skipDirectoryCheck
         };
 
         const attributes = ZosFilesAttributes.loadFromFile(
@@ -52,7 +54,6 @@ export default class FileToUSSHandler extends ZosFilesBaseHandler {
         if (attributes != null) {
             uploadOptions.attributes = attributes;
         }
-
         const response = await Upload.uploadFile(
             session,
             commandParameters.arguments.inputfile,